+++ team5/src/main/java/team5/model/Timeslot.java
@@ -5,11 +5,12 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.time.LocalDateTime;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class Timeslot {
 
-    private static final AtomicInteger count = new AtomicInteger(0);
-    private final int id;
+    private static final AtomicLong count = new AtomicLong(0);
+    private final long id;
     private LocalDateTime startDateTime;
     private LocalDateTime endDateTime;
     private boolean isAvailable;
@@ -25,7 +26,7 @@ public class Timeslot {
         this.id = count.incrementAndGet();
     }
 
-    public int getId() { return id; }
+    public long getId() { return id; }
 
     public LocalDateTime getStartDateTime() {
         return startDateTime;
